+++ packages/tools/src/isIn.ts
@@ -1,8 +1,10 @@
-export const isIn = (
-  dom: HTMLElement,
-  classNames: string | string[],
-  parent = document.body,
-) => {
+/**
+ * check is `dom` inside parent with `classNames`
+ * @param dom
+ * @param classNames
+ * @param parent
+ */
+export default function isIn(dom: HTMLElement, classNames: string | string[], parent = document.body) {
   let cName = classNames;
   if (typeof classNames === 'string') {
     cName = [classNames];
@@ -20,4 +22,21 @@ export const isIn = (
     if (dom.parentElement) return isIn(dom.parentElement, classNames, parent);
   }
   return false;
-};
+}
+
+/**
+ * check is `dom` inside something
+ * @param dom
+ * @param callback
+ * @param parent
+ */
+export function isInBy(dom: HTMLElement, callback: (node: HTMLElement) => boolean, parent = document.body) {
+  if (dom) {
+    if (dom === parent) {
+      return false;
+    }
+    if (callback(dom)) return true;
+    if (dom.parentElement) return isInBy(dom.parentElement, callback, parent);
+  }
+  return false;
+}