+++ packages/pocket/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @orca-fe/pocket
 
+## 3.0.1
+
+### Patch Changes
+
+- update dependencies
+
 ## 3.0.0
 
 ### Major Changes